openai/requirements.txt
fastapi
uvicorn
PyPDF2
scikit-learn
google-generativeai
requests
python-dotenv
